==> 02_scraping_equipos.ipynb <==
import numpy as np
import pandas as pd
import os

import requests
from  bs4 import BeautifulSoup

from selenium import webdriver
from selenium.webdriver.common.by import By

from time import sleep
import random

browser = webdriver.Chrome()

browser.get("https://nba.com/games")
browser.maximize_window()

# cookies
sleep(1)
browser.find_element(by=By.ID, value="onetrust-accept-btn-handler").click()

# Ir a la página de equipos
sleep(2)
browser.find_element(by=By.XPATH, value='//*[@id="nav-ul"]/li[7]/a').click()


html_equipos = browser.page_source
soup = BeautifulSoup(html_equipos, "html.parser")
urls_equipos_nba = [a['href'] for a in soup.find_all('a', class_='Anchor_anchor__cSc3P TeamFigureLink_teamFigureLink__uqnNO') if '/stats/team/' in a['href']]
print(urls_equipos_nba)

# Para acceder a las estadísticas del primer equipo
# NO EJECUTAR si queremos hacer directamente el bucle con todos
sleep(0.3)

browser.find_element(By.CSS_SELECTOR, '#__next > div.Layout_base__6IeUC.Layout_justNav__2H4H0 > div.Layout_mainContent__jXliI > div.MaxWidthContainer_mwc__ID5AG > section > div > div.TeamDivisions_wrapper__5_SVo > div:nth-child(1) > div.TeamDivisions_divisionTeams__N6vcW > div:nth-child(1) > div > div > a:nth-child(2)').click()


# NO EJECUTAR si queremos hacer el bucle con todos, esto solo hace una extracción que me sirvió de guía
# Igual que hicimos con la de jugador, sacamos la ID del equipo del html
url_actual = browser.current_url
id_equipo = url_actual.split('/')[-1]

# Obtener el nombre del equipo
nombre_equipo_elemento = browser.find_element(By.CSS_SELECTOR, ".TeamHeader_name__MmHlP")
nombre_equipo = nombre_equipo_elemento.text

# Esto es para obtener la racha de victorias y también la posición en la división
registro_elemento = browser.find_element(By.CSS_SELECTOR, ".TeamHeader_record__wzofp")
registro_datos = registro_elemento.text.split('|')
victorias_derrotas = registro_datos[0].strip()
puesto = registro_datos[1].split('in')[0].strip()
division = registro_datos[1].split('in')[1].strip()

# Esto es para obtener las estadísticas esas del cuadro superior derecho de la página de equipos
ppg_elemento = browser.find_element(By.CSS_SELECTOR, "div.TeamHeader_rank__lMnzF:nth-of-type(1) .TeamHeader_rankValue__ZGDCq")
ppg = ppg_elemento.text

rpg_elemento = browser.find_element(By.CSS_SELECTOR, "div.TeamHeader_rank__lMnzF:nth-of-type(2) .TeamHeader_rankValue__ZGDCq")
rpg = rpg_elemento.text

apg_elemento = browser.find_element(By.CSS_SELECTOR, "div.TeamHeader_rank__lMnzF:nth-of-type(3) .TeamHeader_rankValue__ZGDCq")
apg = apg_elemento.text

oppg_elemento = browser.find_element(By.CSS_SELECTOR, "div.TeamHeader_rank__lMnzF:nth-of-type(4) .TeamHeader_rankValue__ZGDCq")
oppg = oppg_elemento.text


print(f"ID del equipo: {id_equipo}")
print(f"Nombre del equipo: {nombre_equipo}")
print(f"Registro (Victorias-Derrotas): {victorias_derrotas}")
print(f"Puesto: {puesto}, División: {division}")
print(f"PPG: {ppg}, RPG: {rpg}, APG: {apg}, OPPG: {oppg}")


def extraer_informacion_equipo_bs4(url):
    try:        
        response = requests.get(url)
        soup = BeautifulSoup(response.text, 'html.parser')
        
        id_equipo = url.split('/')[-1]
        
        nombre_equipo_elemento = soup.select_one(".TeamHeader_name__MmHlP")
        nombre_equipo = nombre_equipo_elemento.text if nombre_equipo_elemento else "none"

        registro_elemento = soup.select_one(".TeamHeader_record__wzofp")
        if registro_elemento:
            registro_datos = registro_elemento.text.split('|')
            victorias_derrotas = registro_datos[0].strip()
            puesto = registro_datos[1].split('in')[0].strip()
            division = registro_datos[1].split('in')[1].strip()
        else:
            victorias_derrotas = puesto = division = "none"

        ppg = rpg = apg = oppg = "none"
        estadisticas_elementos = soup.select("div.TeamHeader_rank__lMnzF .TeamHeader_rankValue__ZGDCq")
        if estadisticas_elementos:
            ppg = estadisticas_elementos[0].text if len(estadisticas_elementos) > 0 else "none"
            rpg = estadisticas_elementos[1].text if len(estadisticas_elementos) > 1 else "none"
            apg = estadisticas_elementos[2].text if len(estadisticas_elementos) > 2 else "none"
            oppg = estadisticas_elementos[3].text if len(estadisticas_elementos) > 3 else "none"

        return {
            "id_equipo": id_equipo,
            "nombre_equipo": nombre_equipo,
            "victorias_derrotas": victorias_derrotas,
            "puesto": puesto,
            "division": division,
            "ppg": ppg,
            "rpg": rpg,
            "apg": apg,
            "oppg": oppg,
        }
    except Exception as e:
        print(f"Error al acceder: {e}")
        return None


id_equipo, nombre_equipo, victorias_derrotas, puesto, division, ppg, rpg, apg, oppg = [], [], [], [], [], [], [], [], []


url_base = "https://www.nba.com"
web_equipos_nba = [url_base + div.select("a")[1]["href"] for div in soup.find_all("div", class_="TeamFigure_tfLinks__gwWFj")]

for url in web_equipos_nba:
    print(f"Accediendo a: {url}")
    info_equipo = extraer_informacion_equipo_bs4(url)
    if info_equipo:
        id_equipo.append(info_equipo['id_equipo'])
        nombre_equipo.append(info_equipo['nombre_equipo'])
        victorias_derrotas.append(info_equipo['victorias_derrotas'])
        puesto.append(info_equipo['puesto'])
        division.append(info_equipo['division'])
        ppg.append(info_equipo['ppg'])
        rpg.append(info_equipo['rpg'])
        apg.append(info_equipo['apg'])
        oppg.append(info_equipo['oppg'])

    
    sleep(random.uniform(2, 4))


df_equipos = pd.DataFrame({
    'ID Equipo': id_equipo,
    'Nombre Equipo': nombre_equipo,
    'Victorias-Derrotas': victorias_derrotas,
    'Puesto': puesto,
    'División': division,
    'PPG': ppg,
    'RPG': rpg,
    'APG': apg,
    'OPPG': oppg
})


print(df_equipos)

df_equipos.to_excel("excels/equipos_nba.xlsx", index=False, engine='openpyxl')


# Hacemos el guardado que estamos ya realizando para si existse el archivo, que no se sobreescriba

def guardar_excel_con_numeracion(df, nombre_base):
    directorio = os.path.join(os.getcwd(), "excels")

    if not os.path.exists(directorio):
        os.makedirs(directorio)

    ruta_base = os.path.join(directorio, nombre_base)

    contador = 0
    ruta_final = f"{ruta_base}.xlsx"

    while os.path.exists(ruta_final):
        contador += 1
        ruta_final = f"{ruta_base}_{contador}.xlsx"

    df.to_excel(ruta_final, index=False, engine='openpyxl')
    print(f"Archivo guardado como: {ruta_final}")


guardar_excel_con_numeracion(df_equipos, "equipos_nba")
