==> 06_limpiar_datos_pandas.ipynb <==
# Primero ejecutar el archivo 05_limpiar_datos.pandas.ipynb antes que este. Es importante!!!

import pandas as pd
import os
import re


archivo_1 = 'excels/jugadores_nba.xlsx'
archivo_2 = 'excels/actualizados/estadistica_jugadores_nba_actualizado.xlsx'

df1 = pd.read_excel(archivo_1)
df2 = pd.read_excel(archivo_2)

print(df1.columns)
print(df2.columns)

df1.drop('Equipo', axis=1, inplace=True)
df2.drop(['Player'], axis=1, inplace=True)

# Vamos a limpiar la columna dorsal. Vamos a quitar las almohadillas y si hay algún valor que no sea numérico o está en blanco ponerle none

df1['Dorsal'] = df1['Dorsal'].str.replace('#', '', regex=True)
df1['Dorsal'] = df1['Dorsal'].apply(lambda x: None if not str(x).isdigit() else x)
print(df1['Dorsal'])

conteo_posiciones = df1['Posición'].value_counts()
valores_blancos = df1['Posición'].isna().sum()
print("Conteo de Posiciones:\n", conteo_posiciones)
print("\nCantidad de valores en blanco:", valores_blancos)

# Voy a cambiar el nombre de las posiciones a español
# Primero creamos un diccionario
posiciones_español = {
    'Guard': 'Base',
    'Forward': 'Alero',
    'Center': 'Pivot',
    'Forward-Center': 'Ala-Pivot',
    'Guard-Forward': 'Base - Alero',
    'Forward-Guard': 'Alero-Escolta',
    'Center-Forward': 'Ala-Pivot'
}

df1['Posición'] = df1['Posición'].replace(posiciones_español)

# Y ahora en los sitios que esté blanco o no haya datos se le pone none en texto
df1['Posición'] = df1['Posición'].apply(lambda x: 'none' if x not in posiciones_español.values() else x)
print(df1['Posición'])

# Ahora vamos a trabajar a la vez con las cuatro columnas numéricas de datos del dataframe

columnas = ['PPG', 'RPG', 'APG', 'PIE']

# La idea es que si no tiene número que me ponga none en texto
def valores_numero(valor):
    try:
        return pd.to_numeric(valor)
    except ValueError:
        return 'none'

for columna in columnas:
    df1[columna] = df1[columna].apply(valores_numero)

print(df1['PPG'])
print(df1['RPG'])
print(df1['APG'])
print(df1['PIE'])

# Vamos ahora a trabjar con la columna altura. Lo primero es cambiarle el nombre para que ponga Altura (metros) para saber ese dato
# Y luego la idea es dejar solo el valor en metros del dato. Utilizando expresiones regulares

df1.rename(columns={'Altura': 'Altura (metros)'}, inplace=True)

def extraer_altura_metros(valor):
    if pd.isna(valor):
        return None
    try:
        # Utilizamos una expresión regular
        altura_metros = pd.Series(valor).str.extract(r'\((\d+\.\d+)m\)')[0]
        return altura_metros
    except Exception:
        return 'none'


df1['Altura (metros)'] = df1['Altura (metros)'].apply(extraer_altura_metros)
print(df1['Altura (metros)'])


import pandas as pd

# Cambiar el nombre de la columna 'Peso' a 'Peso (kg)'
df1.rename(columns={'Peso': 'Peso (kg)'}, inplace=True)

# Función para extraer el valor en kilogramos
def extraer_peso_kg(valor):
    if pd.isna(valor):
        return 'none'
    try:
        # Extraer la parte en kilogramos utilizando una expresión regular
        peso_kg = pd.Series(valor).str.extract(r'\((\d+)kg\)')[0]
        return peso_kg
    except Exception:
        return 'none'

# Aplicar la función a la columna 'Peso (kg)'
df1['Peso (kg)'] = df1['Peso (kg)'].apply(extraer_peso_kg)

print(df1['Peso (kg)'])

# Función para limpiar la columna Edad
def limpiar_edad(valor):
    if pd.isna(valor) or not str(valor).endswith("years"):
        return None
    return str(valor).replace("years", "").strip()

# Aplicar la función a la columna Edad
df1['Edad'] = df1['Edad'].apply(limpiar_edad)

# Función para limpiar la columna Experiencia
def limpiar_experiencia(valor):
    if pd.isna(valor):
        return None
    if valor == "Rookie":
        return valor
    if str(valor).lower().endswith("years") or str(valor).lower().endswith("year"):
        return str(valor).replace("Years", "").replace("years", "").replace("Year", "").replace("year", "").strip()
    return None

# Aplicar la función a la columna Experiencia
df1['Experiencia'] = df1['Experiencia'].apply(limpiar_experiencia)

print(df1['Edad'])
print(df1['Experiencia'])

# Ahora vamos con la columna cumpleaños, para ponerla en formato fecha más reconocible en España

df1['Cumpleaños'] = pd.to_datetime(df1['Cumpleaños'], errors='coerce').dt.strftime('%d-%m-%Y')

df1['Cumpleaños'] = df1['Cumpleaños'].where(df1['Cumpleaños'].notna(), None)


print(df1['Cumpleaños'])


# Fusionar df2 con df1 usando 'id_jugador' como clave, preservando todos los registros de df2
df_jugadores = pd.merge(df2, df1, on='id_jugador', how='left', suffixes=('', '_df1'))

# Lista del orden deseado de columnas
orden_columnas = ['id_jugador', 'Nombre', 'Apellido', 'ID Equipo'] + \
                 [col for col in df1.columns if col not in ['id_jugador', 'Nombre', 'Apellido']] + \
                 [col for col in df2.columns if col not in ['id_jugador', 'ID Equipo']]

# Reordenar las columnas en df_jugadores
df_jugadores = df_jugadores[orden_columnas]

# Rellenar con None los valores NaN resultantes de la fusión
df_jugadores.fillna('None', inplace=True)



ruta_archivo = 'excels/actualizados/jugadores_completos.xlsx'

df_jugadores.to_excel(ruta_archivo, index=False)
