==> 07_conectar_airtable.ipynb <==
import numpy as np
import pandas as pd
import os
import requests
from pathlib import Path

from pprint import pprint

API_KEY = "xxxxx" # Usuario

BASE_ID = "xxxxxx" # Base: Ensaladas - API

TABLE_ID = "xxxx" # Tabla: datos_1

airtable_base_url = "https://api.airtable.com/v0"

# Headers
headers = {"Authorization" : f"Bearer {API_KEY}",
           "Content-Type"  : "application/json"}

pprint(headers)

# Endpoint
endpoint = f"{airtable_base_url}/{BASE_ID}/{TABLE_ID}"

print(endpoint)

params = {"fields"                : None, 
          "maxRecords"            : None, 
          "pageSize"              : None,
          "returnFieldsByFieldId" : None}

print(endpoint)

pprint(params, sort_dicts = False)

archivo_excel = 'excels/actualizados/equipos_nba_actualizado.xlsx'
df1 = pd.read_excel(archivo_excel)

df1 = df1.replace({np.nan: None})

datos_subir = {
    "records": [{"fields": df1.iloc[i, :].to_dict()} for i in range(min(10, len(df1)))],
    "typecast": True
}
pprint(datos_subir)

def enviar_a_airtable(grupo_datos):
    response = requests.post(url=endpoint, json=grupo_datos, headers=headers)
    # Aquí puedes añadir manejo de la respuesta para verificar si fue exitosa

# Bucle para enviar los datos en grupos de 10
for start in range(0, len(df1), 10):
    end = min(start + 10, len(df1))  # Asegura no sobrepasar el límite del DataFrame
    grupo_datos = {
        "records": [{"fields": row.to_dict()} for _, row in df1.iloc[start:end].iterrows()],
        "typecast": True
    }
    enviar_a_airtable(grupo_datos)

# Ahora vamos a hacer lo mismo que lo anterior, para enviar la tabla de estadistícas de los jugadores
# Hay que volver a poner credenciales porque la tabla es diferente

API_KEY = "xxxxxxx" # Usuario

BASE_ID = "xxxxxxx"  

TABLE_ID = "xxxxxx" 

airtable_base_url = "https://api.airtable.com/v0"

# Headers
headers = {"Authorization" : f"Bearer {API_KEY}",
           "Content-Type"  : "application/json"}

pprint(headers)

# Endpoint
endpoint = f"{airtable_base_url}/{BASE_ID}/{TABLE_ID}"

print(endpoint)

params = {"fields"                : None, 
          "maxRecords"            : None, 
          "pageSize"              : None,
          "returnFieldsByFieldId" : None}

print(endpoint)

pprint(params, sort_dicts = False)

archivo_excel = 'excels/actualizados/jugadores_completos.xlsx'
df2 = pd.read_excel(archivo_excel)

df2 = df2.replace({np.nan: None})

datos_subir_jug = {
    "records": [{"fields": df2.iloc[i, :].to_dict()} for i in range(min(10, len(df2)))],
    "typecast": True
}
pprint(datos_subir_jug)

def enviar_a_airtable(grupo_datos):
    response = requests.post(url=endpoint, json=grupo_datos, headers=headers)
  
# Bucle para enviar los datos de los jugadores en grupos de 10 por la restricción de la API de airtable
for start in range(0, len(df2), 10):
    end = min(start + 10, len(df2))  
    grupo_datos = {
        "records": [{"fields": row.to_dict()} for _, row in df2.iloc[start:end].iterrows()],
        "typecast": True
    }
    enviar_a_airtable(grupo_datos)
