==> 08_descargar_airtable.ipynb <==
import numpy as np
import pandas as pd
import os
import requests
from pathlib import Path

from pprint import pprint

API_KEY = "xxxxxx" # Usuario

BASE_ID = "xxxxxx" # Base del espacio de la NBA

TABLE_ID = "xxxxxxx" # Tabla de los equipos NBa

airtable_base_url = "https://api.airtable.com/v0"

# Headers
headers = {"Authorization" : f"Bearer {API_KEY}",
           "Content-Type"  : "application/json"}

pprint(headers)

# Primero vamos a establecer el endpoint 

endpoint = f"{airtable_base_url}/{BASE_ID}/{TABLE_ID}"
params = {
    "fields": None, 
    "maxRecords": None, 
    "pageSize": None,
    "returnFieldsByFieldId": None
}
print(endpoint)

response = requests.get(url=endpoint, headers=headers, params=params)


datos_equipos_nba = list()
while params.get("offset") != None or len(datos_equipos_nba) == 0:
    response = requests.get(url=endpoint, headers=headers, params=params)
    data = response.json()
    offset = data.get("offset")
    params["offset"] = offset
    datos_equipos_nba.extend(data["records"])

pd.json_normalize(datos_equipos_nba)

# Ahora vamos a hacer lo mismo que lo anterior, para descargar la tabla de estadistícas de los jugadores
# Hay que volver a poner credenciales porque la tabla es diferente

API_KEY = "pat0pGqotA9LgB5u3.1da7001a502355b60289fde3449bf1a129d5b038aa7c7020f9f170e21ec31a49" # Usuario

BASE_ID = "appx9WDAOvTxfMLY2"  # workspasece base

TABLE_ID = "tblgKGDzjTmkl2nve"  # tabla de datos estadísticos jugadores

airtable_base_url = "https://api.airtable.com/v0"

# Headers
headers = {"Authorization" : f"Bearer {API_KEY}",
           "Content-Type"  : "application/json"}

pprint(headers)

# Endpoint
endpoint = f"{airtable_base_url}/{BASE_ID}/{TABLE_ID}"

print(endpoint)

params = {"fields"                : None, 
          "maxRecords"            : None, 
          "pageSize"              : None,
          "returnFieldsByFieldId" : None
          }

response = requests.get(url=endpoint, headers=headers, params=params)



datos_jugadores_nba = list()
while params.get("offset") != None or len(datos_jugadores_nba) == 0:
    response = requests.get(url=endpoint, headers=headers, params=params)
    data = response.json()
    offset = data.get("offset")
    params["offset"] = offset
    datos_jugadores_nba.extend(data["records"])

pd.json_normalize(datos_jugadores_nba)

df_datos_equipos_nba = pd.DataFrame(datos_equipos_nba)
df_datos_jugadores_nba = pd.DataFrame(datos_jugadores_nba)

# Para trabajar con los datos descargados en un nuevo nootbook vamos a guardar los datos en excel
# Además crearé la ruta por si al ejecutar alguien no lo tiene que se guarde ahí

ruta_guardado = "excels/descargados/"
if not os.path.exists(ruta_guardado):
    os.makedirs(ruta_guardado)

archivo_1 = os.path.join(ruta_guardado, "datos_equipos_nba.xlsx")
archivo_2 = os.path.join(ruta_guardado, "datos_jugadores_nba.xlsx")

df_datos_equipos_nba.to_excel(archivo_1, index=False)
df_datos_jugadores_nba.to_excel(archivo_2, index=False)

# Me he dado cuenta que al pasarlo así se me queda una cosa muy fea, en formato AIRTABLE y no me gusta para hacer gráficas
# ni para leer los datos. Así que voy a trabajar con pandas para hacer el dataframe a mi gusto
# Voy a decirle que todas los datos del diccionario en el fields sean columnas y luego almacenar los datos
# Además eliminaré la columna id y la de tiempo que no me valen para nada para hacer gráficas

#Tenemos ya los datos en dataframe
df_datos_equipos_nba
df_datos_jugadores_nba

# Por lo que he visto en la documentación lo primero que hay qeu hacer es utilizar normalize a los fields
normalizar_equipos = pd.json_normalize(df_datos_equipos_nba['fields'])
normalizar_jugadores = pd.json_normalize(df_datos_jugadores_nba['fields'])

# Aquí para que lo tengamos en cuenta me volví loco hasta que encontré el fallo, había que poner una axis
# porque si no hacía las cosas en las filas y no en la columna... 
df_bueno_equipos = pd.concat([df_datos_equipos_nba.drop('fields', axis=1), normalizar_equipos], axis=1)
df_bueno_jugadores = pd.concat([df_datos_jugadores_nba.drop('fields', axis=1), normalizar_jugadores], axis=1)




# Y ahora, antes de guardar, vamos a eliminar las dos primeras columnas con pandas que es la id de airtable y la de tiempo que no las queremos para nada
df_bueno_equipos = df_bueno_equipos.drop(['id', 'createdTime'], axis=1)
df_bueno_jugadores = df_bueno_jugadores.drop(['id', 'createdTime'], axis=1)


archivo_1 = os.path.join(ruta_guardado, "datos_equipos_nba.xlsx")
archivo_2 = os.path.join(ruta_guardado, "datos_jugadores_nba.xlsx")

df_bueno_equipos.to_excel(archivo_1, index=False)
df_bueno_jugadores.to_excel(archivo_2, index=False)
